+++ engine/src/agent/human_web_agent.cpp
@@ -165,7 +165,7 @@ MoveDecision HumanWebAgent::make_move(const MoveRequest& req) {
       if (type == "move") {
         auto it = obj.find("index");
         if (it != obj.end() && it->value().is_int64()) {
-          long idx = long(it->value().as_int64());
+          long idx = it->value().as_int64();
           if (idx >= 0 && size_t(idx) < plays.size()) {
             return plays[size_t(idx)];
           }
