+++ engine/src/training/move_set_eval_target_log.cpp
@@ -70,7 +70,7 @@ void TargetWriter::add_position(uint32_t game_index, uint32_t turn_index,
   TargetPositionHeader ph{};
   ph.game_index = game_index;
   ph.turn_index = turn_index;
-  ph.num_candidates = uint32_t(candidates.size());
+  ph.num_candidates = candidates.size();
   ph.num_legal_moves = num_legal_moves;
   append_bytes(&buffer_, &ph, sizeof(ph));
   std::array<uint8_t, kPlaneCells> quantized;
